refactor: modularization of utils
